feat(glushkov): adjust base project to the new construction
