+++ src/lib.rs
@@ -1,2 +1,3 @@
-mod dfa;
+mod glushkov;
 pub mod regex_engine;
+mod thompson;
